fixed file load sequential
